moved the cable stock move to the correct model
